Fix minor typo in Thinking In Nushell section
The book referred to `save` as a file, saying "you pipe to a file that has the job of saving content". This is confusing and potentially misleading - files don't have any job. Instead, we pipe to a command that has the job of saving content and give that command a path argument for the file that the content should be saved to.