Fix minor typos and clarify `Parsing and evaluation...` section (#329)
* Fix minor typo in Thinking In Nushell section

The book referred to `save` as a file, saying "you pipe to a file that has the job of saving content". This is confusing and potentially misleading - files don't have any job. Instead, we pipe to a command that has the job of saving content and give that command a path argument for the file that the content should be saved to.

* Clarify `Parsing and evaluation...` Section

I think this is by far the most confusing section for many people coming to Nu. I know there have been discussions of how to clarify it on Discord, but I figured I'd submit some minor changes as a first effort.

Personally, I think that the given example is more confusing than is necessary. This is especially true because if you run each line as a separate command in the shell, it will work perfectly - and this is how most people will think to try the script. 

Instead, I would highlight the second example and use that to explain the overall concept. It's much more clear (to me, at least) that sourcing a dynamic path doesn't work because the source command tries to parse the contents of the path before the dynamic expression has been evaluated. This example has the added benefit that it would fail even if run from the shell. We could, for example, give this self-contained example:

```bash
> source $nu.config-path
```

* Fix minor typo in Introduction