FIX: update the links to the default config files
In this commit, i've `grep`ed all the files containing either
(1) "/default_config.nu" or (2) "/default_env.nu".
The links did point to either
https://github.com/nushell/nushell/blob/main/docs/sample_config/default_*.nu
or
https://github.com/nushell/nushell/blob/main/src/sample_config/default_*.nu
However, these files do not exist anymore from recent commits and
the links are broken in the book itself...

I double checked and the only location where one can find the
default config files is now in
https://github.com/nushell/nushell/blob/main/crates/nu-utils/src/sample_config/default_*.nu

This commit replaces all occurences of the old and broken links to
their new locations in nushell/nushell.