[pt-BR] fix statement about `lines` output type
which is a `list`, not a `table`.

This should fix it in the Brazilian Portuguese translation.
See PR #764 for a fix of the same problem
in the English original and in the German translation.

I don't really know Portuguese. I simply looked up the term to use in
https://www.nushell.sh/pt-BR/book/tipos_de_dados.html#listas
(and in https://www.nushell.sh/pt-BR/book/tipos_de_dados.html#dados-estruturados
where it occurs in singular, albeit not referring to the data type),
so it's quite possible I'm introducing grammatical errors.