[es] fix statement about `lines` output type (#765)
which is a `list`, not a `table`.

This should fix it in the Spanish translation.
See PR #764 for a fix of the same problem
in the English original and in the German translation.

I don't really know Spanish. I simply looked up the term to use in
https://www.nushell.sh/es/book/tipos_de_datos.html#listas
(and in https://www.nushell.sh/es/book/tipos_de_datos.html#datos-estructurados
where it occurs in singular, albeit not referring to the data type),
so it's quite possible I'm introducing grammatical errors.