Fix custom completion scripts (#959)
* Fix fish completer returning empty string

* Fix completer override with fallback example

* Fix fish completer edge cases

Wgeb the first element of the completion list doesn't have a
description, `from tsv` throws an error. When using the `--flexible`
flag, the second column disappears.

This solution adds headers to the text, that way we'll get the two
columns. As a bonus, we avoid having to rename the columns :)