Automate response to PRs trying to update commands (#969)
* Automate response to PRs trying to update commands

We frequently see PRs that try to fix or improve command docs. As the
pages are autogenerated this is a futile endeavour.

We regularly have to respond to them and close the PR.

Attempt to provide a workflow that will react to PRs that affect the
command docs path and send a comment.
I don't autoclose for now, as the autogenerated still come in as a PR.

* Try to tweak body for correct markdown

* Resort to basic js