Update command signature doc to modern syntax (#1167)
* Update command signature doc to modern syntax

While implementing a custom alias, I found myself unable to provide an explicit command signature as described in the documentation. A bit of crawling the main repo's commit history led to my discovery that the syntax as described in the documentation was indeed outdated. After a bit of experimentation I was able to discover the correct annotation syntax.
Ideally, in a slightly longer timescale than I am willing to deal with right now, this change shouldn't be needed; I believe the information on this page would serve better integrated with the rest of the documentation. It seems as though it would fit better there after the syntax change.

* Update signature doc with single type example

This should bring the syntax more in-line with the discussion in #1035.

* Update signature doc with clearer examples.

I've also removed the `cell-path` paragraph. It's probably fine(?) as again that should already be elsewhere.
Furthermore, I've decided that instead of omitting some types from `str distance` that instead it would be better to just select a command with one non-`<nothing>` pipeline signature. Surprisingly challenging! `histogram` seems like it should do the trick, though.

* Update signature doc primary example

amtoine suggested that `str stats` is used in the first example rather than `histogram`, as unlike `str stats`, `histogram` is not a default command.