Replace Pipelines "Behind the Scenes" with "Result Display Rendering"
The previous "Behind the Scenes" docs has some problems:

* The title is not descriptive
* Confusing context reference: The "You may have wondered" references `ls` output despite no applicable example in front of it - and is not stable against other docs section changes either way
* Confusing false equivalence: Claims "one and the same" and then explains how they are not in fact the same

The section is being replaced with a "Result Display Rendering", which

* Mentions and explains the existence of a display output hook
* Begins with establishing context: The section is about interactive shell pipeline result display output

---

I kept the section small as not to duplicate content from the Hooks documentation page.

Adding examples for current, different, or simplified displaying may be reasonable here, but would duplicate the hooks "Changing how Output is Displayed" and "Filtering or Diverting Command Output" sections.