Update and simplify Escaping to the System Windows Note
The history note (historic behavior, in the past) is very old. It has little value in the current docs.

The "shell built-ins" wording and description may not be obvious to readers.

The Coming from CMD.EXE documents the internal command forwarding as well as other relations between the two shells.
Here, in 'Escaping the System', we should focus on the aspect of escaping to the system, only give a general description and note, and reference to the more definitive and exhaustive documentation.