Drop experimental label from display_output hook (#1644)
As per comment https://github.com/nushell/nushell/pull/14361#issuecomment-2483400412
it's not clear what "experimental" would mean, the hook was introduced two years ago and has been stable since,
and in #14361 there is a discussion about the hooks behavior but no indication that it would be dropped.

Dropping the experimental label makes it more obvious that the hook is here to stay. If its behavior will change, it can be documented like usual in release breaking change notes (which would make sense anyway).

As a drive-by text form fixup, drop the period from the last two list items. None of the other list items end in a period, which matches common form for [simple] lists.