Fix hooks.md basic hooks examples and hooks list (#1889)
Fix hooks syntax; they are a list of values. The code as it was did not run.

Nushell has been using an inherent configuration and the user configuration only replacing parts.
For the hooks it makes sense to set only the hooks value.

Replace "you can also use a list", which is the default and used above already, with a "you can append additional hooks" example.

For zh-CN, at least add the missing available hook docs in English instead of missing them altogether.

For zh-CN, add the new example replacement with EN text rather than keeping the old outdated one, or missing it altogether.