Drop examples from closure language reference
Relates to #1929

The language reference should be focused on the language concepts, not describe how various commands make use of it, and especially not in the form of examples.

All these commands have closure examples in the docs comments. All listed commands link to their respective docs.

Dropping the examples increases focus of the *reference* doc, and reduces risks of doc duplication and of command changes leading to outdated lang ref docs.