Reduce and focus closure command examples (#1982)
In this section of the language reference we want to explore and demonstrate examples of variance of use.

The unnecessary layout separation (headlines, horizontal line separators) and command elaboration took away from that.
This change focuses the examples into demonstrating different closure use cases and behaviors in the context of commands.

It also adjusts the examples to have one code block with input and output.

It also improves some code examples, like the sort-by actually doing some sorting, demonstrating.

With the sometimes adjusted examples and documented output, the closure behavior does not need the redundant explanation.