Use more concise home feature title (#1970)
Use `Cross-platform` instead of `Pipelines to control any OS`.

It's not necessarily clear what pipelines refers to. Pipelines may refer to one of many use cases. Pipelines may be a misleading ambiguity to command piping (shell terminology).

The previous title lead to a line break. Making it more concise to fit on one line is more consistent with the other two feature titles - leading to a more consistent and balanced layout too. (Body starts on the same vertical position now.)