Remove mentions of the `chocolatey` package (#2023)
There seems to be an issue with a shimming mechanism interposed by
chocolatey that turns a `Ctrl-C`/`SIGINT` from the controlling terminal
into a killing signal. Thus core functionality seems to be impaired:
- https://github.com/nushell/nushell/issues/6428
- https://github.com/chocolatey/home/issues/134
- There appears to be ways to configure chocolatey packages to not use
this shim. This responsibility would be on the package maintainer
- https://community.chocolatey.org/packages/nushell
- https://github.com/dgalbraith/chocolatey-packages/tree/master/automatic/nushell

Thus we can not recommend this particular package source until the
issues with that has been resolved.